# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MarkReadIn": {
        "properties": {
          "ids": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ids"
          }
        },
        "title": "MarkReadIn",
        "type": "object"
      },
...
      "MessageIn": {
        "properties": {
          "body": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          }
        },
        "required": [
          "body"
        ],
        "title": "MessageIn",
        "type": "object"
      },
      "MessageOut": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "deal_id": {
            "format": "uuid",
            "title": "Deal Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "mine": {
            "default": false,
            "title": "Mine",
            "type": "boolean"
          },
          "sender_name": {
            "title": "Sender Name",
            "type": "string"
          },
          "sender_org_id": {
            "format": "uuid",
            "title": "Sender Org Id",
            "type": "string"
          },
          "sender_user_id": {
            "format": "uuid",
            "title": "Sender User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "deal_id",
          "sender_user_id",
          "sender_name",
          "sender_org_id",
          "body",
          "created_at"
        ],
        "title": "MessageOut",
        "type": "object"
      },
...
      "NotificationPreferenceIn": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "boolean"
          },
          "in_app": {
            "title": "In App",
            "type": "boolean"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "in_app",
          "email"
        ],
        "title": "NotificationPreferenceIn",
        "type": "object"
      },
      "Ok": {
        "properties": {
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "title": "Ok",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Programmable escrow for agentic commerce. The LLM never moves money: it writes a signed attestation, and a deterministic settlement engine reads it.",
    "title": "Aegis",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/chat/deals/{deal_id}": {
      "get": {
        "operationId": "messages_api_v1_chat_deals__deal_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "deal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Deal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MessageOut"
                  },
                  "title": "Response Messages Api V1 Chat Deals  Deal Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Messages",
        "tags": [
          "platform"
        ]
      },
      "post": {
        "description": "Deal-scoped chat.  Never fed to the verifier as evidence.",
        "operationId": "send_message_api_v1_chat_deals__deal_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "deal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Deal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send Message",
        "tags": [
          "platform"
        ]
      }
    },
...
    "/api/v1/notifications/mark-read": {
      "post": {
        "operationId": "mark_read_api_v1_notifications_mark_read_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkReadIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Mark Read Api V1 Notifications Mark Read Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mark Read",
        "tags": [
          "platform"
        ]
      }
    },
    "/api/v1/notifications/preferences": {
      "get": {
        "operationId": "preferences_api_v1_notifications_preferences_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Preferences Api V1 Notifications Preferences Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Preferences",
        "tags": [
          "platform"
        ]
      },
      "put": {
        "operationId": "set_preference_api_v1_notifications_preferences_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationPreferenceIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Preference",
        "tags": [
          "platform"
        ]
      }
    },
...
    "/api/v1/payments/deals/{deal_id}/payouts": {
      "get": {
        "operationId": "payouts_api_v1_payments_deals__deal_id__payouts_get",
        "parameters": [
          {
            "in": "path",
            "name": "deal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Deal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Payouts Api V1 Payments Deals  Deal Id  Payouts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Payouts",
        "tags": [
          "platform"
        ]
      }
    },
...
    "/api/v1/realtime/chat/{deal_id}": {
      "get": {
        "operationId": "sse_chat_api_v1_realtime_chat__deal_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "deal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Deal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sse Chat",
        "tags": [
          "platform"
        ]
      }
    },
...
    "/api/v1/realtime/verification/{milestone_id}": {
      "get": {
        "operationId": "sse_verification_api_v1_realtime_verification__milestone_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "milestone_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Milestone Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sse Verification",
        "tags": [
          "platform"
        ]
      }
    },
    "/api/v1/reputation/entities/{entity_id}": {
      "get": {
        "description": "A counterparty passport is intentionally readable across tenants: it is the\npublic trading record the other side needs before agreeing to a deal.  It\ncontains no deal contents, no evidence and no member identities.",
        "operationId": "reputation_api_v1_reputation_entities__entity_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Reputation Api V1 Reputation Entities  Entity Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reputation",
        "tags": [
          "platform"
        ]
      }
    },
    "/api/v1/settlements/deals/{deal_id}": {
      "get": {
        "operationId": "settlements_api_v1_settlements_deals__deal_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "deal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Deal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Settlements Api V1 Settlements Deals  Deal Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Settlements",
        "tags": [
          "platform"
        ]
      }
    },
...
  }
}